# ValidationService/ValidationService/api/ValidationService.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ValidationService/api/ValidationService.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ValidationService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/validate/cart/add": {
      "post": {
        "operationId": "ValidationService_ValidateAddToCart",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/validationGetCartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/validationAddToCartRequest"
            }
          }
        ],
        "tags": [
          "ValidationService"
        ]
      }
    },
    "/validate/cart/{userId}": {
      "get": {
        "operationId": "ValidationService_ValidateGetCart",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/validationGetCartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "ValidationService"
        ]
      },
      "delete": {
        "operationId": "ValidationService_ValidateClearCart",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/validationGetCartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "ValidationService"
        ]
      }
    },
    "/validate/cart/{userId}/item/{skuId}": {
      "delete": {
        "operationId": "ValidationService_ValidateDeleteItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/validationGetCartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "skuId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "ValidationService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "validationAddToCartRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "uint64"
        },
        "skuId": {
          "type": "string",
          "format": "uint64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "validationCartItem": {
      "type": "object",
      "properties": {
        "skuId": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        },
        "price": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "validationGetCartResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/validationCartItem"
          }
        },
        "totalPrice": {
          "type": "string",
          "format": "uint64"
        }
      }
    }
  }
}